+++ resources/js/Pages/Settings/Privacy/Partials/PrivacyRule.vue
